feat: icon 존재 확인 로직 추가
